added disclaimer component
